Hassan changes in misc items

+++ app/Http/Controllers/books_materialController.php
@@ -6,6 +6,9 @@ use App\pages;
 use App\objects;
 use App\states;
 use App\Videos;
+use App\Questions;
+use App\Quiz;
+use App\User;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Input;
 use Illuminate\Http\Request;
@@ -13,13 +16,17 @@ use Illuminate\Support\Facades\View;
 use phpDocumentor\Reflection\DocBlock\Tags\Return_;
 
 use Illuminate\Pagination\LengthAwarePaginator;
+use Auth;
 
 class testit{
     public $page_id = -1;
 }
-class books_materialController extends Controller
+class books_materialController extends commonController
 {
+  public function __construct(){
 
+    parent::__construct();
+  }
 
     public function addCategory()
     {
@@ -62,7 +69,7 @@ class books_materialController extends Controller
 
     public function showBooksList()
     {
-        $books = DB::table('books')->get();
+        $books = DB::table('books')->where('user_id',@$this->data->cUserId)->get();
         //return view('books/Books', array("data" => $books));
         return view('books/Books1', array("data" => $books));
     }
@@ -97,13 +104,14 @@ class books_materialController extends Controller
         $input['title'] = Input::get('title');
         $input['description'] = Input::get('description');
         $input['category_id'] =Input::get('category_id');
+        $input['user_id'] =@$this->data->cUserId;
         books::create($input);
         return redirect('Books');
     }
 
     public function getBookRecord($id)
     {
-        $book=DB::table('books')->where('id',$id)->first();
+        $book=DB::table('books')->where('id',$id)->where('user_id',@$this->data->cUserId)->first();
         return view('books/EditBook' , array("data"=>$book));
     }
 
@@ -126,6 +134,7 @@ class books_materialController extends Controller
                     'title'=> Input::get('title'),
                     'description'=> Input::get('description'),
                     'category_id'=>Input::get('category_id'),
+
                     ]);
         return redirect('Books');
     }
@@ -872,7 +881,7 @@ class books_materialController extends Controller
 
     public function showVideoList(){
 
-        $videos = DB::table('videos')->simplePaginate(3);
+        $videos = DB::table('videos')->where('user_id',$this->data->cUserId)->simplePaginate(5);
 
         return view('videos/videos', array("data" => $videos));
 
@@ -907,13 +916,14 @@ class books_materialController extends Controller
         $input['url']=$url=Input::get('url');
         $input['description']=Input::get('description');
         $input['thumbnail']=$video_obj->tokenize($url);
+         $input['user_id'] = $this->data->cUserId;
         Videos::create($input);
 
 
     }
     public function deleteVideo($id){
 
-        DB::table('videos')->where('id', $id)->delete();
+        DB::table('videos')->where('user_id',$this->data->cUserId)->where('id', $id)->delete();
         return redirect('videos');
 
     }
@@ -956,7 +966,32 @@ class books_materialController extends Controller
         return view('videos/detail',array('data'=>$datail));
     }
 
+ public function dashboard(){
 
+
+   if(@$this->data->cUserType == 'admin'){
+        @$this->data->category = categories::all()->count();
+             @$this->data->books = books::all()->count();
+             @$this->data->videos = Videos::all()->count();
+             @$this->data->questions = Questions::all()->count();
+             @$this->data->quiz = Quiz::all()->count();
+             @$this->data->users = User::all()->count();
+           $data = (array)$this->data;
+                return view('dashboards/admin', compact('data'));
+   }
+   elseif (@$this->data->cUserType == 'author') {
+   
+             @$this->data->books = books::all()->where('user_id',$this->data->cUserId)->count();
+            @$this->data->videos = Videos::all()->count();
+             @$this->data->questions = Questions::all()->count();
+             @$this->data->quiz = Quiz::all()->where('user_id',$this->data->cUserId)->count();
+            
+           $data = (array)$this->data;
+      $data = (array)$this->data;
+                return view('dashboards/author', compact('data'));
+   }
+     
+    }
 
 
 }
